requirements/dev.txt
discord.py
black
flake8
pillow
requests
aiohttp
aiosqlite
python-dotenv